+++ app/src/main/java/ru/plamit/currencytest/entity/CurrencyRates.kt
@@ -1,11 +1,12 @@
 package ru.plamit.currencytest.entity
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class CurrencyRates(@SerializedName("date")
                          val date: String = "",
                          @SerializedName("rates")
-                         val rates: Map<String, Double>,
+                         val rates: Map<String, BigDecimal>,
                          @SerializedName("base")
                          val base: String = "")
